+++ stats/opentelemetry/experimental/grpc_trace_bin_propagator.go
@@ -16,7 +16,7 @@
  *
  */
 
-package opentelemetry
+package experimental
 
 import (
 	"context"
